build: client and interaction structures
